LU-19180 osc: split i/o for servers without unaligned_dio
When sending RPCs to older older servers without unaligned_dio
in cases where an RPC would be large and unaligned the request
can be failed -EINVAL.

However this is sub-optimal.

Instead the error should be unique to the failure and the caller
can check for the error, -EDEADLK, and split the RPC request
performing a short i/o and re-queuing the remaining pages.

Test-Parameters: serverversion=2.15
Fixes: 262957eaef ("LU-18691 quota: quota interop check for 64k page clients")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I63985ad380439cf9bbebdd7c3ef41fa96e3dfc63